fix: resolve some issues
- Resolve incorrect name max stale
- Fix incorrect refresh media stale logic
- Use last timestamp every update metadata
