+++ core/models/src/main/kotlin/com/waffiq/bazz_movies/core/models/Favorite.kt
@@ -26,9 +26,11 @@ data class Favorite(
 
   @Suppress("MagicNumber")
   fun isStale(): Boolean {
-    val ninetyDaysMillis =
+    // tmdb terms says max data cached is 90 days,
+    // but we use earlier (10 days before the date),
+    val eightyDaysMillis =
       80L * 24 * 60 * 60 * 1000
 
-    return System.currentTimeMillis() - lastUpdated > ninetyDaysMillis
+    return System.currentTimeMillis() - lastUpdated > eightyDaysMillis
   }
 }
